+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 
 ### Additions
 
+- There is no longer a need to set the Modem hex address in config.yaml.
+  ([PR 303][P303])
+
 ### Fixes
 
 - Fix error with backlight setting (thanks @tommycw1)([PR 299][P299])
@@ -550,3 +553,4 @@ will add new features.
 [I292]: https://github.com/TD22057/insteon-mqtt/issues/292
 [P285]: https://github.com/TD22057/insteon-mqtt/pull/285
 [P299]: https://github.com/TD22057/insteon-mqtt/pull/299
+[P303]: https://github.com/TD22057/insteon-mqtt/pull/303